detached CameraShakeController
