+++ Assets/Scripts/General/Controllers/CameraShakeController.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 23bdc95e0064439a982473f8acdf8b78
+timeCreated: 1612804911
